fix rubocop issues, move courtfinder services under c100_app
