+++ app/services/c100_app/postcode_screen_decision_tree.rb
@@ -23,7 +23,7 @@ module C100App
       end
 
     # CourtPostcodeChecker already logs the exception
-    rescue
+    rescue StandardError
       show(:error_but_continue)
     end
   end
