+++ app/views/steps/postcode_screen/error_but_continue/show.html.erb
@@ -11,7 +11,6 @@
     <!-- TODO: Add the copy needed, and adapt the button linking to another step -->
 
     <div class="xform-group form-submit">
-      <%= link_to t('.go_back'), edit_steps_postcode_screen_children_postcodes_path, class: 'link-back link-button' %>
       <%= link_to t('.continue'), '/steps/miam/consent_order', class: 'button', role: 'button' %>
     </div>
   </div>
