+++ spec/services/c100_app/postcode_screen_decision_tree_spec.rb
@@ -16,21 +16,21 @@ RSpec.describe C100App::PostcodeScreenDecisionTree do
   
     context 'and no valid courts are found' do
       before do
-        allow_any_instance_of(CourtPostcodeChecker).to receive(:courts_for).and_return([])
+        allow_any_instance_of(C100App::CourtPostcodeChecker).to receive(:courts_for).and_return([])
       end
       it { is_expected.to have_destination(:no_court_found, :show) }
     end
   
     context 'and at least one valid court is found' do
       before do
-        allow_any_instance_of(CourtPostcodeChecker).to receive(:courts_for).and_return(['i am a court'])
+        allow_any_instance_of(C100App::CourtPostcodeChecker).to receive(:courts_for).and_return(['i am a court'])
       end
       it { is_expected.to have_destination('/steps/miam/consent_order', :edit) }
     end
 
     context 'when the postcode checker raises an error' do
       before do
-        allow_any_instance_of(CourtPostcodeChecker).to receive(:courts_for).and_raise("expected exception for testing, please ignore")
+        allow_any_instance_of(C100App::CourtPostcodeChecker).to receive(:courts_for).and_raise("expected exception for testing, please ignore")
       end
       it { is_expected.to have_destination(:error_but_continue, :show)}
     end
